MultiCI: Abstract Github Action state reading
This adds a new pattern for abstracting away the "github action"-specific code within the github actions. The
overall intention is to perform this change for all of the `@actions/xxx` libraries in use to enable
implementation of these features within other environments (e.g. Jenkins, BuildKite, GitLab).

Part of gradle/actions#502.