MultiCI: Abstract Logging behind new type
In order to support more environments, we abstract away `@actions/core`'s logging functionality
behind it's own type. This allows replacement with environment-specific versions.

Note: This PR does not replace all, there is a follow-up to finish replacing. The intention is
to have the new content reviewed vs. the "find and replace" PR following.

Part of gradle/actions#502.